feat: 通过 store 控制提示
